moved ConnectionManager and ExecutionException
