+++ src/com/clouway/jdbc/ExecutionException.java
@@ -1,4 +1,4 @@
-package com.clouway.jdbc.travel.agency;
+package com.clouway.jdbc;
 
 /**
  * @author Krasimir Raikov(raikov.krasimir@gmail.com)
